Gestion d'accessibilité des pages en fonction de l'utilisateur et affichage de l'auteur

+++ delete-article.php
@@ -1,6 +1,17 @@
 <?php
 
-$pdo = require './database/database.php';
+require_once __DIR__ . '/database/database.php';
+require_once __DIR__ . '/database/security.php';
+
+$currentUser = isLoggedIn();
+
+if (!$currentUser) {
+    header('Location: /');
+}
+
+/**
+ * @var ArticleDAO
+ */
 
 $articleDAO = require './database/models/ArticleDAO.php';
 
@@ -8,6 +19,15 @@ $_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
 $id = $_GET['id'] ?? '';
 
 if ($id) {
-    $articleDAO->deleteOne($id);
+
+    $article = $articleDAO->getOne($id);
+
+    if ($currentUser['id'] === $article['author']) {
+        $articleDAO->deleteOne($id);
+    } else {
+        header('Location :/');
+    }
+
 }
+
 header('Location: /');
